break work resume out into component, bring Employer component inside, add react-router-dom
