+++ package.json
@@ -9,6 +9,7 @@
     "react-addons-css-transition-group": "^15.6.2",
     "react-dom": "^16.4.2",
     "react-motion": "^0.5.2",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.4"
   },
   "scripts": {
